Overview
Large amount of data is being generated daily, however only data by itself is not useful. The insights taken from it are the important information you gather by analysing the data.

Data visualization allows to represent these insights in a visual context made by graphs, charts and etc, which is a technique that is very useful for the audiences comprehend.

Besides gather information to discover insights and display in graphical formats which is a very effective way to show information and communication. You have to know your audience, as different generations responds and interact in distinct ways on digital platforms.

The objective of this assignment is to analyse and create interactive data visualization from a data set “consolidated_coin_data.csv” and as a result, use the visual representation to communicate the information for two different groups: 17-35 years old and 60+ age group. We are looking for to answer these two question bellow:

1- Details change in the “High” value for all the Currencies over time.

2- Depicting the “Volume” by Currency between 2016 and 2019 for the top 3 Currencies.

To answer the questions and display in an interactive visualization for both generations mentioned previously, first of all we need to prepare the data.

This assignment will be using python programming language. In addition to it pandas, which is a very useful library to work with data because it provides tools to read data from various formats and transform into dataframes which makes easier to handle the data.

After that I will be using plotly to create all the interactive visualization for both audience. Different from pandas where it is only possible to create static plots, with plotly you can create interactive visualization. Since the main concepts are: data, layout and figure. So that it allows you to reuse the code and build graphs for both generations creatively in way to achieve the aesthetic appeal for each user group.